// src/types/place.ts
export interface Place {
	id: string
	title: string
	type: string
	price: number
	city: {
		name: string
		location: {
			latitude: number
			longitude: number
			zoom: number
		}
	}
	location: {
		latitude: number
		longitude: number
		zoom: number
	}
	isFavorite: boolean
	isPremium: boolean
	rating: number
	previewImage: string
}

export type PlaceCard = Omit<Place, 'city' | 'location'>